Refactor: Split `aggregate_dyod.cpp/.hpp` into separate files
