+++ src/lib/operators/aggregate/aggregate_vector.hpp
@@ -1,0 +1,124 @@
+#pragma once
+
+#include <algorithm>
+#include <unordered_set>
+#include <vector>
+
+#include "types.hpp"
+#include "window_function_traits.hpp"
+
+namespace hyrise {
+
+class AbstractAggregateVector {
+ public:
+  virtual ~AbstractAggregateVector() = default;
+  virtual void grow_if_necessary(size_t size) = 0;
+  virtual void merge(AbstractAggregateVector& other) = 0;
+
+  size_t count(GroupID group_id) {
+    return _counts[group_id];
+  }
+
+  void increment_count(GroupID group_id) {
+    _counts[group_id]++;
+  }
+
+  std::vector<size_t>& counts() {
+    return _counts;
+  }
+
+  GroupID group_count() const {
+    return _group_count;
+  }
+
+ protected:
+  size_t _group_count;
+  std::vector<size_t> _counts;
+};
+
+template <typename ColumnDataType, WindowFunction aggregate_function>
+struct TypedAggregateVector : AbstractAggregateVector {
+  using AggregateDataType = typename WindowFunctionTraits<ColumnDataType, aggregate_function>::ReturnType;
+  using DistinctValues = std::unordered_set<ColumnDataType>;
+  using AccumulatorDataType =
+      std::conditional_t<aggregate_function == WindowFunction::CountDistinct, DistinctValues, AggregateDataType>;
+
+ public:
+  // The mutable accessor is needed because aggregator functions mutate values directly.
+  AccumulatorDataType& accumulator(GroupID group_id) {
+    return _accumulators[group_id];
+  }
+
+  std::vector<AccumulatorDataType>& accumulators() {
+    return _accumulators;
+  }
+
+  void grow_if_necessary(size_t size) override {
+    if (_counts.size() < size) {
+      _counts.resize(size);
+      _accumulators.resize(size);
+    }
+  }
+
+  void merge(AbstractAggregateVector& other) override {
+    auto& typed_other = static_cast<TypedAggregateVector<ColumnDataType, aggregate_function>&>(other);
+    _merge(typed_other);
+  }
+
+ protected:
+  std::vector<AccumulatorDataType> _accumulators;
+
+  void _merge(TypedAggregateVector<ColumnDataType, aggregate_function>& other) {
+    const auto new_size = std::max(_accumulators.size(), other._accumulators.size());
+
+    _accumulators.resize(new_size);
+    _counts.resize(new_size);
+
+    if constexpr (aggregate_function == WindowFunction::CountDistinct) {
+      auto& other_accumulators = other._accumulators;
+      const auto other_size = other_accumulators.size();
+
+      for (auto index = size_t{0}; index < new_size; ++index) {
+        if (index < other_size) {
+          _accumulators[index].merge(other_accumulators[index]);
+        }
+      }
+    } else if constexpr (aggregate_function == WindowFunction::Count) {
+      const auto& other_counts = other._counts;
+      const auto other_size = other_counts.size();
+
+      for (auto index = size_t{0}; index < new_size; ++index) {
+        if (index < other_size && other_counts[index] > 0) {
+          _counts[index] += other_counts[index];
+        }
+      }
+    } else if constexpr (aggregate_function == WindowFunction::Any) {
+      const auto& other_accumulators = other._accumulators;
+      const auto& other_counts = other._counts;
+      const auto other_size = other_accumulators.size();
+
+      for (auto index = size_t{0}; index < new_size; ++index) {
+        if (_counts[index] == 0 && index < other_size && other_counts[index] > 0) {
+          _accumulators[index] = other_accumulators[index];
+          _counts[index] += other_counts[index];
+        }
+      }
+    } else {
+      const auto& other_accumulators = other._accumulators;
+      const auto& other_counts = other._counts;
+      const auto other_size = other_accumulators.size();
+
+      auto aggregator =
+          WindowFunctionBuilder<AggregateDataType, AggregateDataType, aggregate_function>().get_aggregate_function();
+
+      for (auto index = size_t{0}; index < new_size; ++index) {
+        if (index < other_size && other_counts[index] > 0) {
+          aggregator(other_accumulators[index], _counts[index], _accumulators[index]);
+          _counts[index] += other_counts[index];
+        }
+      }
+    }
+  }
+};
+
+}  // namespace hyrise
